Made clippy happy and ran rustfmt.

+++ src/lib.rs
@@ -568,7 +568,6 @@ impl<S: Storage + Send + Sync + 'static> SelfEncryptor<S> {
     }
 
     /// Truncate the self_encryptor to the specified size (if extended, filled with 0u8).
-    //
     // NOTE: Right now calls prepare_window with the full length of the truncated file
     // because chunk sizes may change. Chunks that don't change don't actually need any
     // treatment here so there's room for performance improvement, especially for large
@@ -586,8 +585,7 @@ impl<S: Storage + Send + Sync + 'static> SelfEncryptor<S> {
             let num_chunks = self.get_num_chunks(); // call with updated self.file_size
             self.chunks.truncate(num_chunks as usize);
         } else {
-            if new_size > MAX_IN_MEMORY_SIZE as u64 &&
-               self.sequencer.len() <= MAX_IN_MEMORY_SIZE {
+            if new_size > MAX_IN_MEMORY_SIZE as u64 && self.sequencer.len() <= MAX_IN_MEMORY_SIZE {
                 match self.sequencer.create_mapping() {
                     Ok(()) => (),
                     Err(_) => return false,
@@ -1344,18 +1342,18 @@ mod test {
     fn truncate_three_max_chunks() {
         let storage = Arc::new(SimpleStorage::new());
         let data_map: DataMap;
-        let bytes_len = MAX_CHUNK_SIZE as u64 * 3;
+        let bytes_len = MAX_CHUNK_SIZE * 3;
         let bytes = random_bytes(bytes_len as usize);
         {
             let mut se = SelfEncryptor::new(storage.clone(), DataMap::None);
             se.write(&bytes, 0);
-            check_file_size(&se, bytes_len);
-            se.truncate(bytes_len - 24);
+            check_file_size(&se, bytes_len as u64);
+            se.truncate(bytes_len as u64 - 24);
             assert_eq!(se.get_num_chunks(), 3);
-            check_file_size(&se, bytes_len - 24);
+            check_file_size(&se, bytes_len as u64 - 24);
             data_map = se.close();
         }
-        assert_eq!(data_map.len(), bytes_len - 24);
+        assert_eq!(data_map.len(), bytes_len as u64 - 24);
         match data_map {
             DataMap::Chunks(ref chunks) => {
                 assert_eq!(chunks.len(), 3);
@@ -1367,30 +1365,30 @@ mod test {
             _ => panic!("datamap should be DataMap::Chunks"),
         }
         let mut se = SelfEncryptor::new(storage.clone(), data_map);
-        let fetched = se.read(0, bytes_len - 24);
-        assert!(&fetched[..] == &bytes[.. (bytes_len - 24) as usize]);
+        let fetched = se.read(0, bytes_len as u64 - 24);
+        assert!(&fetched[..] == &bytes[..(bytes_len - 24) as usize]);
     }
 
     #[test]
     fn truncate_from_datamap() {
         let storage = Arc::new(SimpleStorage::new());
-        let bytes_len = MAX_CHUNK_SIZE as u64 * 3;
+        let bytes_len = MAX_CHUNK_SIZE * 3;
         let bytes = random_bytes(bytes_len as usize);
         let data_map: DataMap;
         {
             let mut se = SelfEncryptor::new(storage.clone(), DataMap::None);
             se.write(&bytes, 0);
-            check_file_size(&se, bytes_len);
+            check_file_size(&se, bytes_len as u64);
             data_map = se.close();
         }
         let data_map2: DataMap;
         {
             // Start with an existing datamap.
             let mut se = SelfEncryptor::new(storage.clone(), data_map);
-            se.truncate(bytes_len - 24);
+            se.truncate(bytes_len as u64 - 24);
             data_map2 = se.close();
         }
-        assert_eq!(data_map2.len(), bytes_len - 24);
+        assert_eq!(data_map2.len(), bytes_len as u64 - 24);
         match data_map2 {
             DataMap::Chunks(ref chunks) => {
                 assert_eq!(chunks.len(), 3);
@@ -1402,30 +1400,30 @@ mod test {
             _ => panic!("datamap should be DataMap::Chunks"),
         }
         let mut se = SelfEncryptor::new(storage.clone(), data_map2);
-        let fetched = se.read(0, bytes_len - 24);
-        assert!(&fetched[..] == &bytes[.. (bytes_len - 24) as usize]);
+        let fetched = se.read(0, bytes_len as u64 - 24);
+        assert!(&fetched[..] == &bytes[..(bytes_len - 24) as usize]);
     }
 
     #[test]
     fn truncate_to_extend_from_datamap() {
         let storage = Arc::new(SimpleStorage::new());
-        let bytes_len = MAX_CHUNK_SIZE as u64 * 3 - 24;
+        let bytes_len = MAX_CHUNK_SIZE * 3 - 24;
         let bytes = random_bytes(bytes_len as usize);
         let data_map: DataMap;
         {
             let mut se = SelfEncryptor::new(storage.clone(), DataMap::None);
             se.write(&bytes, 0);
-            check_file_size(&se, bytes_len);
+            check_file_size(&se, bytes_len as u64);
             data_map = se.close();
         }
         let data_map2: DataMap;
         {
             // Start with an existing datamap.
             let mut se = SelfEncryptor::new(storage.clone(), data_map);
-            se.truncate(bytes_len + 24);
+            se.truncate(bytes_len as u64 + 24);
             data_map2 = se.close();
         }
-        assert_eq!(data_map2.len(), bytes_len + 24);
+        assert_eq!(data_map2.len(), bytes_len as u64 + 24);
         match data_map2 {
             DataMap::Chunks(ref chunks) => {
                 assert_eq!(chunks.len(), 3);
@@ -1437,9 +1435,9 @@ mod test {
             _ => panic!("datamap should be DataMap::Chunks"),
         }
         let mut se = SelfEncryptor::new(storage.clone(), data_map2);
-        let fetched = se.read(0, bytes_len + 24);
-        assert!(&fetched[.. bytes_len as usize] == &bytes[..]);
-        assert!(&fetched[bytes_len as usize ..] == &[0u8; 24]);
+        let fetched = se.read(0, bytes_len as u64 + 24);
+        assert!(&fetched[..bytes_len as usize] == &bytes[..]);
+        assert!(&fetched[bytes_len as usize..] == &[0u8; 24]);
     }
 
     #[test]
